Вынес виджеты в отдельные файлы
