+++ lib/presentation/todos_list/todo_list.dart
@@ -1,0 +1,23 @@
+import 'package:flutter/material.dart';
+import 'package:todos/domain/models/todo.dart';
+import 'package:todos/presentation/todos_list/todo_card.dart';
+
+class TodoList extends StatelessWidget {
+  final List<Todo> todos;
+
+  TodoList(this.todos);
+
+  @override
+  Widget build(BuildContext context) {
+    const kEmptySpaceForFabHeight = 88.0;
+
+    return todos.isEmpty
+        ? const Center(child: Text('Нет элементов'))
+        : ListView(
+            children: [
+              ...todos.map((e) => TodoCard(e)).toList(),
+              const SizedBox(height: kEmptySpaceForFabHeight),
+            ],
+          );
+  }
+}
