+++ lib/routes.dart
@@ -1,12 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:todos/data/repositories/mock_todos_repository.dart';
-import 'package:todos/presentation/todos_list/todos_list_screen.dart';
+import 'package:todos/presentation/todos_list/todo_list_screen.dart';
 
-const initialRoute = TodosListScreen.routeName;
+const initialRoute = TodoListScreen.routeName;
 
 final routes = <String, WidgetBuilder>{
-  TodosListScreen.routeName: (context) =>
-      TodosListScreen(MockTodosRepository()),
+  TodoListScreen.routeName: (context) => TodoListScreen(MockTodosRepository()),
 };
 
 final RouteFactory onGenerateRoute = null;
